[NEW] Payment service

+++ src/main/java/io/abun/wmb/TransactionService/dto/payment/PaymentDetailRequest.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record PaymentDetailRequest(
         @JsonProperty("gross_amount")
-        Long amount,
+        Integer amount,
 
         @JsonProperty("order_id")
         String orderId
